Change FIELD_* MD5s and move interlaced files to conformance/passed/
FFmpeg correctly decodes all samples of interlaced content, however
outputs them a field at a time, as opposed to a frame at a time like
VTM.  The result is a sequence with half the height and twice the rate
of those produced by VTM.  One can use the weave filter to perform a
simple deinterlace and produce identical output to VTM.  The MD5s of the
correctly decoded content used in conformance testing expect the decoder
to output interlaced content a frame at a time like VTM.

The fact FFmpeg outputs interlaced raw video in this way is not
indicative of any defect in the decoder, rather it is quite an arbitrary
choice.  We could choose to selectively apply the weave filter for only
the FIELD_* bitstreams. This could be done by manually configuring the
command line options for these bitstreams, or using a custom app which
has this behaviour.  The simpler option taken here however, is to modify
the MD5s to match those of the field-at-a-time output produced by the
ffmpeg CLI by default.